Search
Added a search button and search box that will allow a user to search for any NASDAQ stock symbol.

When a stock is selected, it's name, stock symbol, price and date for this query will be displayed.

A user can also close the selected stock information by clicking the close button.

Formatting for the price might need to be taken into account and different decimal placements have shown up.